refactored the custom hooks to start with a lowercase letter instead of capital letter
